+++ Tests/WindowsCloudPC.Tests.ps1
@@ -13,6 +13,7 @@ BeforeDiscovery {
         'Get-CloudPCUserSetting',
         'Invoke-CloudPCPolicyReprovision',
         'Invoke-CloudPCReprovision',
+        'New-CloudPCSnapshot',
         'Restart-CloudPC'
     )
 }